ASYLUM: Cleanup Pass 3
- Add proper init code for Object & BlowUpPuzzle classes
- Use Common::Rational in a few more places
- Add casts & const modifiers in proper places

git-svn-id: http://asylumengine.googlecode.com/svn/trunk@547 0bfb4aae-4ea4-11de-8d8d-752d95cf3e3c

+++ engines/asylum/system/screen.cpp
@@ -127,9 +127,9 @@ void Screen::setPalette(byte *rgbPalette) const {
 	p += 3;
 
 	for (int32 i = 1; i < 256; i++) {
-		palette[i * 4 + 0] = *p++ << 2;
-		palette[i * 4 + 1] = *p++ << 2;
-		palette[i * 4 + 2] = *p++ << 2;
+		palette[i * 4 + 0] = (byte)(*p++ << 2);
+		palette[i * 4 + 1] = (byte)(*p++ << 2);
+		palette[i * 4 + 2] = (byte)(*p++ << 2);
 		palette[i * 4 + 3] = 0;
 	}
